+++ CHANGELOG.md
@@ -20,6 +20,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Close file lists in migration ([#1191](https://github.com/scm-manager/scm-manager/pull/1191))
 - Use command in javahg.py from registrar (Upgrade to newer javahg version)  ([#1192](https://github.com/scm-manager/scm-manager/pull/1192))
 - Fixed wrong e-tag format ([sdorra/web-resource #1](https://github.com/sdorra/web-resources/pull/1))
+- Handles repositories in custom directories correctly in migration fron 1.x ([#1201](https://github.com/scm-manager/scm-manager/pull/1201))
 
 ## [2.0.0] - 2020-06-04
 ### Added